Adds attributes to XML for international labels.
This commit adds to the `XML` used to purchase postage, so international
labels can be purchased through the gem. To use it one must include a
hash with a `customs` key. Within this hash needs to be certain
attributes. I have included a sample.
```
{
   ContentsType: 'Merchandise', #This is the type of contents.
   CustomsCertify: TRUE, #This is either TRUE or FALSE.
   CustomsSigner: 'Jim Johnson', #This is the name of the certifier.
   ToCountryCode: "GB", #Two character country code.
   CustomsItems: [{
     Description: 'Electric Parts Tubes', #Description of the items being sent.
     Quantity: 6, #Quantity of items.
     Weight: 2.5, #Weight in ounces
     Value: 13.86, #Value of items being sent.
     CountryOfOrigin: 'US' #Country items were made.
   }]
}
```
The `CustomsItems` is expected to be an array even if its length is one.
as to any other questions about the particular attributes one can check
the [Endicia Api Docs](https://www.endicia.com/developer/docs/v8.9.html#endicia-els-api),
and search the attributes, because the naming used corresponds with the
`XML` attributes, that are needed, within the documentation.