Fixes issues with non-utf8 characters.
This commit adjusts the parsing of the response from the _Endicia_ api.
Previously, the label was decoded within the gem, but this has been
changed, so the returned label remains base64 encoded. This will allow
the `TubeDepot/tubedepot-spree2` application to save it in the database
as a base64 encoded string. This string will then be send to the print
server where it will be decoded and printed. This was done to help with
labels that contain odd or foreign characters that are difficult to
encoded as utf-8.